Fix for potential nil strings during -drawRect: on iOS 6.
This never gave me fits on iOS 5, but if there was ever an instance
where an FIIconView was being rendered without an icon string set, the
call to create the NSAttributedString "attrString" would fail (as
iconString would still be nil).

By setting iconString to a blank string if iconString was a falsy value
(as in, being nil), everything else ends up going smoothly in drawRect:.
I'm not sure why this was an issue for me in iOS 6 but not in iOS 5
(perhaps it's due to UIKit changes in NSMutableAttributedString?), but I
never even saw a blank icon using this fix.